Change style

+++ src/container/Foods/Foods.tsx
@@ -7,18 +7,25 @@ import AlertMeal from "../../components/AlertMeal/AlertMeal";
 
 const Foods: React.FC = () => {
   const [meals, setMeals] = useState<Meals[]>([]);
+  const [totalPrice, setTotalPrice] = useState<number>(0);
   const [lauding, setLauding] = useState(false);
   const [isActive, setIsActive] = useState(false);
+  const [btnId, setBtnId] = useState<string>("");
 
   const fetchMealsApi = useCallback(async () => {
     try {
       setLauding(true);
       const {data: response} = await axiosApi.get<MealApi | null>(".json");
       if (response) {
-        setMeals(Object.keys(response).map((id) => ({
+        const meals = Object.keys(response).map((id) => ({
           ...response[id],
           id,
-        })));
+        })).reverse();
+        setMeals(meals);
+
+        setTotalPrice(meals.reduce((sum, meal) => {
+          return sum + parseInt(meal.calories);
+        }, 0));
       } else {
         setMeals([]);
       }
@@ -33,8 +40,11 @@ const Foods: React.FC = () => {
 
   const deleteMeal = async (id: string) => {
     setIsActive(true);
-    await axiosApi.delete(`/${id}.json`);
-    void fetchMealsApi();
+    setBtnId(id);
+    if (confirm("Real?")) {
+      await axiosApi.delete(`/${id}.json`);
+      void fetchMealsApi();
+    }
     setIsActive(false);
   };
 
@@ -42,15 +52,28 @@ const Foods: React.FC = () => {
 
   return (lauding) ? <Spinner/> : (
     <div className={"d-flex flex-column"}>
-      <Link to={`/add`} className={`btn btn-secondary ms-auto ${isDisabled}`}>Add new meal</Link>
-      <div>
-        <h3 className={"mb-4"}>Foods</h3>
+      <div className={"d-flex align-items-center"}>
+        {(meals.length)
+          ? <h3 className={"text-success"}>
+            Total calories:{" "}
+            <span className={"text-primary"}>
+            {totalPrice}
+              <strong className={"text-danger"}> kcal</strong>
+          </span>
+          </h3>
+          : undefined
+        }
+        <Link to={`/add`} className={`ms-auto btn btn-success ${isDisabled}`}>Add new meal</Link>
+      </div>
+      <div className={"mt-3"}>
+        <h3 className={"mb-4 text-primary"}>Foods</h3>
         <div>
           {meals.map((meal) => {
             return (
               <AlertMeal
                 key={meal.id}
                 meal={meal}
+                btnId={btnId}
                 isDisabled={isDisabled}
                 isActive={isActive}
                 deleteMeal={deleteMeal}
